refactor(gov) fix(ProposalEntity): removed unnecessary indexes
